MARK FAILURE: Add comments to mark segfault sources
This commit marks a couple of lines that are causing segfaults in
Pytorch with Python 3.13. I'm a bit confused, because we do seem to have
Python 2.9 installed, which should be compatible with Python 3.13.
However, the segfaults I get are on these lines specifically, which both
involve iterating over a loaded dataset.

If we can get these lines to *not* cause a segfault, then this commit
should be reverted.